Use tool name constants and u64 tokens
Add canonical tool-name constants (crates/cinch-rs/src/tools/names.rs) and replace scattered string literals across tools, DAG, filters, file tracking, eviction, and reflection to avoid magic names. Upgrade token counters from u32 to u64 across checkpoint, events, session, harness, execution and sub-agent code paths and switch harness bookkeeping to use the central cost_tracker totals. Introduce a configurable sequential_policy on HarnessConfig (with setter and default), and use it when annotating/injecting tool-call dependencies. Wrap tool errors in structured reflection for LLM self-correction and apply various minor formatting/test cleanups and refactors (truncation helpers, function signatures, added exports in tools/mod and prelude). Removed the AdaptiveRoundLimit implementation from checkpoint.rs. These changes improve robustness for large token counts, make tool renames simpler, and add configurability for execution ordering.

+++ crates/cinch-rs/src/agent/events.rs
@@ -697,9 +697,9 @@ pub struct HarnessResult {
     /// All URL citation annotations collected across rounds.
     pub annotations: Vec<crate::Annotation>,
     /// Total prompt tokens consumed across all rounds.
-    pub total_prompt_tokens: u32,
+    pub total_prompt_tokens: u64,
     /// Total completion tokens consumed across all rounds.
-    pub total_completion_tokens: u32,
+    pub total_completion_tokens: u64,
     /// Number of rounds executed.
     pub rounds_used: u32,
     /// Whether the agent finished naturally (vs hitting the round limit).
@@ -718,7 +718,7 @@ impl HarnessResult {
     }
 
     /// Total tokens (prompt + completion).
-    pub fn total_tokens(&self) -> u32 {
+    pub fn total_tokens(&self) -> u64 {
         self.total_prompt_tokens + self.total_completion_tokens
     }
 }
